Automatically Publish Javadoc to GitHub

+++ test/after_success.sh
@@ -7,24 +7,20 @@ echo "AUTHOR_NAME is $AUTHOR_NAME"
   		  
 if [ "$TRAVIS_REPO_SLUG" == 'NoraUi/NoraUi' ] && [ "$TRAVIS_BRANCH" == 'master' ] && [ "$TRAVIS_PULL_REQUEST" == 'false' ]; then
     
-    echo "******** publish javadoc"
-    cd $HOME
-    git config --global user.email "sgrillon@hotmail.fr"
-    git config --global user.name "Travis-CI"
-    git clone --quiet --branch=master https://${GITHUB_TOKEN}@github.com/NoraUi/NoraUi NoraUi > /dev/null
-
-    # Commit and Push the Changes
-    cd NoraUi
-    git rm -rf ./docs
-    pwd
-    ls -l
-    
-    cp -Rf /home/travis/build/NoraUi/NoraUi/target/apidocs ./docs
-    ls -l
-    
-    git add -f .
-    git commit -m "Lastest javadoc on successful travis build $TRAVIS_BUILD_NUMBER auto-pushed to NoraUi"
-    git push -fq origin master
+    if [ "$AUTHOR_NAME" != 'Travis-CI' ]; then
+        echo "******** publish javadoc"
+        cd $HOME
+        git config --global user.email "sgrillon@hotmail.fr"
+        git config --global user.name "Travis-CI"
+        git clone --quiet --branch=master https://${GITHUB_TOKEN}@github.com/NoraUi/NoraUi NoraUi > /dev/null
+        # Commit and Push the Changes
+        cd NoraUi
+        git rm -rf ./docs
+        cp -Rf /home/travis/build/NoraUi/NoraUi/target/apidocs ./docs
+        git add -f .
+        git commit -m "Lastest javadoc on successful travis build $TRAVIS_BUILD_NUMBER auto-pushed to NoraUi"
+        git push -fq origin master
+    fi
     
     mvn_version=$(mvn -q -Dexec.executable="echo" -Dexec.args='${project.version}' --non-recursive org.codehaus.mojo:exec-maven-plugin:1.5.0:exec)
     echo "Maven version is $mvn_version"
